+++ tests/testthat/test_VCF_variant_caller_unknown.R
@@ -6,14 +6,14 @@ test_that("Test processing VCF with unknown variant caller", {
   file <- "testdata/SBS.GRCh37.variantcaller.unknown.vcf"
   
   # Do not merge adjacent SBSs into DBS
-  list.of.vcfs1 <- ReadAndSplitVCFs(file)
+  list.of.vcfs1 <- ReadAndSplitVCFs(file, filter.status = "PASS")
   
   # Merge adjacent SBSs into DBS
   list.of.vcfs2 <- ReadAndSplitVCFs(file, get.vaf.function = function(x){
     x$VAF <- 0.5
     x$read.depth <- NA
     return(x)
-  })
+  }, filter.status = "PASS")
   expect_equal(nrow(list.of.vcfs1$DBS[[1]]), 0)
   expect_equal(nrow(list.of.vcfs2$DBS[[1]]), 18)
 })